Convert array elements to int and count to size_t to facilitate newP sign check

+++ inc/heurQuickSelect.hpp
@@ -10,11 +10,11 @@
 /* Struct to store local process' data */
 struct localDataHeurQuick
 {
-    uint32_t localMin;
-    uint32_t localMax;
-    uint32_t count = 0;
-    uint32_t leftMargin = 0;
-    uint32_t rightMargin = 0;
+    int localMin;
+    int localMax;
+    size_t count = 0;
+    size_t leftMargin = 0;
+    size_t rightMargin = 0;
 };
 
 /* Find the local process' min and max
@@ -22,7 +22,7 @@ struct localDataHeurQuick
  *      local (output): local process' data whose min and max have been filled
  *      arr (input): the process' array to be searched
  */
-void findLocalMinMax(localDataHeurQuick &local, const std::vector<uint32_t> &arr);
+void findLocalMinMax(localDataHeurQuick &local, const std::vector<int> &arr);
 
 /* Partition the local process' array based on the pivot and find the count
  * @param:
@@ -32,10 +32,10 @@ void findLocalMinMax(localDataHeurQuick &local, const std::vector<uint32_t> &arr
  *      end (input): the last index of the array to search to
  *      p (input): the pivot
  */
-void heurlocalSorting(localDataHeurQuick &local, std::vector<uint32_t> &arr, const size_t start, const size_t end, const uint32_t p);
+void heurlocalSorting(localDataHeurQuick &local, std::vector<int> &arr, const size_t start, const size_t end, const int p);
 
 /* The parallel version of heurlocalSorting */
-void heurParSorting(localDataHeurQuick &local, std::vector<uint32_t> &arr, const size_t start, const size_t end, const uint32_t p);
+void heurParSorting(localDataHeurQuick &local, std::vector<int> &arr, const size_t start, const size_t end, const int p);
 
 /* Set the compare function to use on the elements against the pivot, based on the count of the elements in comparison to k
  * @param:
@@ -43,7 +43,7 @@ void heurParSorting(localDataHeurQuick &local, std::vector<uint32_t> &arr, const
  *      k (input): the index of the sorted array whose value I'm looking for
  *      countSum (input): the number of elements less than or equal to the current pivot
  */
-inline void setComp(bool (*&comp)(const uint32_t &, const uint32_t &), const size_t k, const uint32_t countSum);
+inline void setComp(bool (*&comp)(const int &, const int &), const size_t k, const size_t countSum);
 
 /* Find the closest element to the pivot
  * @param:
@@ -55,7 +55,7 @@ inline void setComp(bool (*&comp)(const uint32_t &, const uint32_t &), const siz
  *      comp (input): the comparison function to compare the elements to the pivot
  *      k (input): the index of the sorted array whose value I'm looking for
  */
-void findClosest(uint32_t &distance, const std::vector<uint32_t> &arr, const size_t start, const size_t end, const uint32_t &p, bool (*comp)(const uint32_t &, const uint32_t &));
+void findClosest(uint32_t &distance, const std::vector<int> &arr, const size_t start, const size_t end, const int &p, bool (*comp)(const int &, const int &));
 
 /* Find the kth element of the array
  * @param:
@@ -65,4 +65,4 @@ void findClosest(uint32_t &distance, const std::vector<uint32_t> &arr, const siz
  *      n (input): the size of the whole array
  *      np (input): the number of processes in the MPI communicator
  */
-void heurQuickSelect(uint32_t &kth, std::vector<uint32_t> &arr, const size_t k, const size_t n, const size_t np);
+void heurQuickSelect(int &kth, std::vector<int> &arr, const size_t k, const size_t n, const size_t np);
