+++ inc/kSearch.hpp
@@ -10,9 +10,9 @@
 /* Struct to store local process' data */
 struct localData
 {
-    uint32_t localMin;
-    uint32_t localMax;
-    uint32_t count = 0;
+    int localMin;
+    int localMax;
+    size_t count = 0;
 };
 
 /* Find the local process' min and max
@@ -20,7 +20,7 @@ struct localData
  *      local (output): local process' data whose min and max have been filled
  *      arr (input): the process' array to be searched
  */
-void findLocalMinMax(localData &local, const std::vector<uint32_t> &arr);
+void findLocalMinMax(localData &local, const std::vector<int> &arr);
 
 /* Find the local process' count
  * @param:
@@ -31,7 +31,7 @@ void findLocalMinMax(localData &local, const std::vector<uint32_t> &arr);
  *      k (input): the index of the sorted array whose value I'm looking for
  *      n (input): the size of the whole array
  */
-inline void findLocalCount(localData &local, const std::vector<uint32_t> &arr, const uint32_t &p, const bool (*comp)(const uint32_t &, const uint32_t &), const size_t k, const size_t n);
+inline void findLocalCount(localData &local, const std::vector<int> &arr, const int &p, const bool (*comp)(const int &, const int &), const size_t k, const size_t n);
 
 /* Find the closest element to the pivot
  * @param:
@@ -40,7 +40,7 @@ inline void findLocalCount(localData &local, const std::vector<uint32_t> &arr, c
  *      p (input): the pivot
  *      comp (input): the comparison function to compare the elements to the pivot
  */
-void findClosest(uint32_t &distance, const std::vector<uint32_t> &arr, const uint32_t &p, const bool (*comp)(const uint32_t &, const uint32_t &));
+void findClosest(uint32_t &distance, const std::vector<int> &arr, const int &p, const bool (*comp)(const int &, const int &));
 
 /* Find the kth element of the array
  * @param:
@@ -50,4 +50,4 @@ void findClosest(uint32_t &distance, const std::vector<uint32_t> &arr, const uin
  *      n (input): the size of the whole array
  *      np (input): the number of processes in the MPI communicator
  */
-void kSearch(uint32_t &kth, std::vector<uint32_t> &arr, const size_t k, const size_t n, const size_t np);
+void kSearch(int &kth, std::vector<int> &arr, const size_t k, const size_t n, const size_t np);
